Added lighting and plane movement
