--- AirplaneGame/Airplane.cs
@@ -1,14 +1,0 @@
-﻿namespace AirplaneGame
-{
-    public class Airplane : Model
-    {
-
-        private const float Gravity = 9.81f;
-        
-        public Airplane(string path) : base(path)
-        {
-        }
-
-
-    }
-}
