+++ AirplaneGame/src/Airplane.cs
@@ -1,0 +1,14 @@
+﻿namespace AirplaneGame
+{
+    public class Airplane : Model
+    {
+
+        private const float Gravity = 9.81f;
+        
+        public Airplane(string path) : base(path)
+        {
+        }
+
+
+    }
+}
